feat: refactor + verify ism factory
