+++ typescript/sdk/src/core/HyperlaneCoreDeployer.ts
@@ -13,11 +13,9 @@ import { HyperlaneDeployer } from '../deploy/HyperlaneDeployer';
 import { ContractVerifier } from '../deploy/verify/ContractVerifier';
 import { HyperlaneHookDeployer } from '../hook/HyperlaneHookDeployer';
 import { HookConfig } from '../hook/types';
-import {
-  HyperlaneIsmFactory,
-  moduleMatchesConfig,
-} from '../ism/HyperlaneIsmFactory';
+import { HyperlaneIsmFactory } from '../ism/HyperlaneIsmFactory';
 import { IsmConfig } from '../ism/types';
+import { moduleMatchesConfig } from '../ism/utils';
 import { MultiProvider } from '../providers/MultiProvider';
 import { ChainMap, ChainName } from '../types';
