+++ typescript/sdk/src/ism/utils.ts
@@ -1,0 +1,415 @@
+import debug from 'debug';
+import { ethers } from 'ethers';
+
+import {
+  DomainRoutingIsm__factory,
+  IAggregationIsm__factory,
+  IInterchainSecurityModule__factory,
+  IMultisigIsm__factory,
+  IRoutingIsm__factory,
+  MailboxClient__factory,
+  OPStackIsm__factory,
+  PausableIsm__factory,
+  StaticAggregationIsm__factory,
+} from '@hyperlane-xyz/core';
+import {
+  Address,
+  eqAddress,
+  formatMessage,
+  normalizeAddress,
+  objMap,
+} from '@hyperlane-xyz/utils';
+
+import { HyperlaneContracts } from '../contracts/types';
+import { ProxyFactoryFactories } from '../deploy/contracts';
+import { MultiProvider } from '../providers/MultiProvider';
+import { ChainName } from '../types';
+
+import {
+  IsmConfig,
+  IsmType,
+  ModuleType,
+  RoutingIsmConfig,
+  RoutingIsmDelta,
+  ismTypeToModuleType,
+} from './types';
+
+const logger = debug('hyperlane:IsmUtils');
+
+// Note that this function may return false negatives, but should
+// not return false positives.
+// This can happen if, for example, the module has sender, recipient, or
+// body specific logic, as the sample message used when querying the ISM
+// sets all of these to zero.
+export async function moduleCanCertainlyVerify(
+  destModule: Address | IsmConfig,
+  multiProvider: MultiProvider,
+  origin: ChainName,
+  destination: ChainName,
+): Promise<boolean> {
+  const originDomainId = multiProvider.tryGetDomainId(origin);
+  const destinationDomainId = multiProvider.tryGetDomainId(destination);
+  if (!originDomainId || !destinationDomainId) {
+    return false;
+  }
+  const message = formatMessage(
+    0,
+    0,
+    originDomainId,
+    ethers.constants.AddressZero,
+    destinationDomainId,
+    ethers.constants.AddressZero,
+    '0x',
+  );
+  const provider = multiProvider.getSignerOrProvider(destination);
+
+  if (typeof destModule === 'string') {
+    const module = IInterchainSecurityModule__factory.connect(
+      destModule,
+      provider,
+    );
+
+    try {
+      const moduleType = await module.moduleType();
+      if (
+        moduleType === ModuleType.MERKLE_ROOT_MULTISIG ||
+        moduleType === ModuleType.MESSAGE_ID_MULTISIG
+      ) {
+        const multisigModule = IMultisigIsm__factory.connect(
+          destModule,
+          provider,
+        );
+
+        const [, threshold] = await multisigModule.validatorsAndThreshold(
+          message,
+        );
+        return threshold > 0;
+      } else if (moduleType === ModuleType.ROUTING) {
+        const routingIsm = IRoutingIsm__factory.connect(destModule, provider);
+        const subModule = await routingIsm.route(message);
+        return moduleCanCertainlyVerify(
+          subModule,
+          multiProvider,
+          origin,
+          destination,
+        );
+      } else if (moduleType === ModuleType.AGGREGATION) {
+        const aggregationIsm = IAggregationIsm__factory.connect(
+          destModule,
+          provider,
+        );
+        const [subModules, threshold] =
+          await aggregationIsm.modulesAndThreshold(message);
+        let verified = 0;
+        for (const subModule of subModules) {
+          const canVerify = await moduleCanCertainlyVerify(
+            subModule,
+            multiProvider,
+            origin,
+            destination,
+          );
+          if (canVerify) {
+            verified += 1;
+          }
+        }
+        return verified >= threshold;
+      } else {
+        throw new Error(`Unsupported module type: ${moduleType}`);
+      }
+    } catch (e) {
+      logger(`Error checking module ${destModule}: ${e}`);
+      return false;
+    }
+  } else {
+    // destModule is an IsmConfig
+    switch (destModule.type) {
+      case IsmType.MERKLE_ROOT_MULTISIG:
+      case IsmType.MESSAGE_ID_MULTISIG:
+        return destModule.threshold > 0;
+      case IsmType.ROUTING: {
+        const checking = moduleCanCertainlyVerify(
+          destModule.domains[destination],
+          multiProvider,
+          origin,
+          destination,
+        );
+        return checking;
+      }
+      case IsmType.AGGREGATION: {
+        let verified = 0;
+        for (const subModule of destModule.modules) {
+          const canVerify = await moduleCanCertainlyVerify(
+            subModule,
+            multiProvider,
+            origin,
+            destination,
+          );
+          if (canVerify) {
+            verified += 1;
+          }
+        }
+        return verified >= destModule.threshold;
+      }
+      case IsmType.OP_STACK:
+        return destModule.nativeBridge !== ethers.constants.AddressZero;
+      case IsmType.TEST_ISM: {
+        return true;
+      }
+      default:
+        throw new Error(`Unsupported module type: ${(destModule as any).type}`);
+    }
+  }
+}
+
+export async function moduleMatchesConfig(
+  chain: ChainName,
+  moduleAddress: Address,
+  config: IsmConfig,
+  multiProvider: MultiProvider,
+  contracts: HyperlaneContracts<ProxyFactoryFactories>,
+  mailbox?: Address,
+): Promise<boolean> {
+  if (typeof config === 'string') {
+    return eqAddress(moduleAddress, config);
+  }
+
+  // If the module address is zero, it can't match any object-based config.
+  // The subsequent check of what moduleType it is will throw, so we fail here.
+  if (eqAddress(moduleAddress, ethers.constants.AddressZero)) {
+    return false;
+  }
+
+  const provider = multiProvider.getProvider(chain);
+  const module = IInterchainSecurityModule__factory.connect(
+    moduleAddress,
+    provider,
+  );
+  const actualType = await module.moduleType();
+  if (actualType !== ismTypeToModuleType(config.type)) return false;
+  let matches = true;
+  switch (config.type) {
+    case IsmType.MERKLE_ROOT_MULTISIG: {
+      // A MerkleRootMultisigIsm matches if validators and threshold match the config
+      const expectedAddress =
+        await contracts.merkleRootMultisigIsmFactory.getAddress(
+          config.validators.sort(),
+          config.threshold,
+        );
+      matches = eqAddress(expectedAddress, module.address);
+      break;
+    }
+    case IsmType.MESSAGE_ID_MULTISIG: {
+      // A MessageIdMultisigIsm matches if validators and threshold match the config
+      const expectedAddress =
+        await contracts.messageIdMultisigIsmFactory.getAddress(
+          config.validators.sort(),
+          config.threshold,
+        );
+      matches = eqAddress(expectedAddress, module.address);
+      break;
+    }
+    case IsmType.FALLBACK_ROUTING:
+    case IsmType.ROUTING: {
+      // A RoutingIsm matches if:
+      //   1. The set of domains in the config equals those on-chain
+      //   2. The modules for each domain match the config
+      // TODO: Check (1)
+      const routingIsm = DomainRoutingIsm__factory.connect(
+        moduleAddress,
+        provider,
+      );
+      // Check that the RoutingISM owner matches the config
+      const owner = await routingIsm.owner();
+      matches &&= eqAddress(owner, config.owner);
+      // check if the mailbox matches the config for fallback routing
+      if (config.type === IsmType.FALLBACK_ROUTING) {
+        const client = MailboxClient__factory.connect(moduleAddress, provider);
+        const mailboxAddress = await client.mailbox();
+        matches =
+          matches &&
+          mailbox !== undefined &&
+          eqAddress(mailboxAddress, mailbox);
+      }
+      const delta = await routingModuleDelta(
+        chain,
+        moduleAddress,
+        config,
+        multiProvider,
+        contracts,
+        mailbox,
+      );
+      matches =
+        matches &&
+        delta.domainsToEnroll.length === 0 &&
+        delta.domainsToUnenroll.length === 0 &&
+        !delta.mailbox &&
+        !delta.owner;
+      break;
+    }
+    case IsmType.AGGREGATION: {
+      // An AggregationIsm matches if:
+      //   1. The threshold matches the config
+      //   2. There is a bijection between on and off-chain configured modules
+      const aggregationIsm = StaticAggregationIsm__factory.connect(
+        moduleAddress,
+        provider,
+      );
+      const [subModules, threshold] = await aggregationIsm.modulesAndThreshold(
+        '0x',
+      );
+      matches &&= threshold === config.threshold;
+      matches &&= subModules.length === config.modules.length;
+
+      const configIndexMatched = new Map();
+      for (const subModule of subModules) {
+        const subModuleMatchesConfig = await Promise.all(
+          config.modules.map((c) =>
+            moduleMatchesConfig(chain, subModule, c, multiProvider, contracts),
+          ),
+        );
+        // The submodule returned by the ISM must match exactly one
+        // entry in the config.
+        const count = subModuleMatchesConfig.filter(Boolean).length;
+        matches &&= count === 1;
+
+        // That entry in the config should not have been matched already.
+        subModuleMatchesConfig.forEach((matched, index) => {
+          if (matched) {
+            matches &&= !configIndexMatched.has(index);
+            configIndexMatched.set(index, true);
+          }
+        });
+      }
+      break;
+    }
+    case IsmType.OP_STACK: {
+      const opStackIsm = OPStackIsm__factory.connect(moduleAddress, provider);
+      const type = await opStackIsm.moduleType();
+      matches &&= type === ModuleType.NULL;
+      break;
+    }
+    case IsmType.TEST_ISM: {
+      // This is just a TestISM
+      matches = true;
+      break;
+    }
+    case IsmType.PAUSABLE: {
+      const pausableIsm = PausableIsm__factory.connect(moduleAddress, provider);
+      const owner = await pausableIsm.owner();
+      matches &&= eqAddress(owner, config.owner);
+
+      if (config.paused) {
+        const isPaused = await pausableIsm.paused();
+        matches &&= config.paused === isPaused;
+      }
+      break;
+    }
+    default: {
+      throw new Error('Unsupported ModuleType');
+    }
+  }
+
+  return matches;
+}
+
+export async function routingModuleDelta(
+  destination: ChainName,
+  moduleAddress: Address,
+  config: RoutingIsmConfig,
+  multiProvider: MultiProvider,
+  contracts: HyperlaneContracts<ProxyFactoryFactories>,
+  mailbox?: Address,
+): Promise<RoutingIsmDelta> {
+  const provider = multiProvider.getProvider(destination);
+  const routingIsm = DomainRoutingIsm__factory.connect(moduleAddress, provider);
+  const owner = await routingIsm.owner();
+  const deployedDomains = (await routingIsm.domains()).map((domain) =>
+    domain.toNumber(),
+  );
+  // config.domains is already filtered to only include domains in the multiprovider
+  const safeConfigDomains = objMap(config.domains, (domain) =>
+    multiProvider.getDomainId(domain),
+  );
+
+  const delta: RoutingIsmDelta = {
+    domainsToUnenroll: [],
+    domainsToEnroll: [],
+  };
+
+  // if owners don't match, we need to transfer ownership
+  if (!eqAddress(owner, normalizeAddress(config.owner)))
+    delta.owner = config.owner;
+  if (config.type === IsmType.FALLBACK_ROUTING) {
+    const client = MailboxClient__factory.connect(moduleAddress, provider);
+    const mailboxAddress = await client.mailbox();
+    if (mailbox && !eqAddress(mailboxAddress, mailbox)) delta.mailbox = mailbox;
+  }
+  // check for exclusion of domains in the config
+  delta.domainsToUnenroll = deployedDomains.filter(
+    (domain) => !Object.values(safeConfigDomains).includes(domain),
+  );
+  // check for inclusion of domains in the config
+  for (const [origin, subConfig] of Object.entries(config.domains)) {
+    const originDomain = safeConfigDomains[origin];
+    if (!deployedDomains.includes(originDomain)) {
+      delta.domainsToEnroll.push(originDomain);
+    } else {
+      const subModule = await routingIsm.module(originDomain);
+      // Recursively check that the submodule for each configured
+      // domain matches the submodule config.
+      const subModuleMatches = await moduleMatchesConfig(
+        destination,
+        subModule,
+        subConfig,
+        multiProvider,
+        contracts,
+        mailbox,
+      );
+      if (!subModuleMatches) delta.domainsToEnroll.push(originDomain);
+    }
+  }
+  return delta;
+}
+
+export function collectValidators(
+  origin: ChainName,
+  config: IsmConfig,
+): Set<string> {
+  // TODO: support address configurations in collectValidators
+  if (typeof config === 'string') {
+    logger.extend(origin)('Address config unimplemented in collectValidators');
+    return new Set([]);
+  }
+
+  let validators: string[] = [];
+  if (
+    config.type === IsmType.MERKLE_ROOT_MULTISIG ||
+    config.type === IsmType.MESSAGE_ID_MULTISIG
+  ) {
+    validators = config.validators;
+  } else if (config.type === IsmType.ROUTING) {
+    if (Object.keys(config.domains).includes(origin)) {
+      const domainValidators = collectValidators(
+        origin,
+        config.domains[origin],
+      );
+      validators = [...domainValidators];
+    }
+  } else if (config.type === IsmType.AGGREGATION) {
+    const aggregatedValidators = config.modules.map((c) =>
+      collectValidators(origin, c),
+    );
+    aggregatedValidators.forEach((set) => {
+      validators = validators.concat([...set]);
+    });
+  } else if (
+    config.type === IsmType.TEST_ISM ||
+    config.type === IsmType.PAUSABLE
+  ) {
+    return new Set([]);
+  } else {
+    throw new Error('Unsupported ModuleType');
+  }
+
+  return new Set(validators);
+}
